Refactor: 회원가입 네비게이션 스택 재구조화
