+++ CheckEat-Business/Views/Auth/Register/RegisterRoute.swift
@@ -1,0 +1,14 @@
+//
+//  RegisterRoute.swift
+//  CheckEat-Business
+//
+//  Created by 최준영 on 8/7/25.
+//
+
+import Foundation
+
+enum RegisterRoute {
+    case businessScan
+    case businessScanResult
+    case registerComplete
+}
